+++ packages/imperative/src/config/__tests__/ProfileInfo.TeamConfig.unit.test.ts
@@ -56,6 +56,8 @@ describe("TeamConfig ProfileInfo tests", () => {
     const teamHomeProjDir = path.join(testDir, testAppNm + "_home_team_config_proj");
     const largeTeamProjDir = path.join(testDir, testAppNm + "_large_team_config_proj");
     const nestedTeamProjDir = path.join(testDir, testAppNm + "_nested_team_config_proj");
+    const nestedTeamProjDirEmptyBase = path.join(testDir, testAppNm + "_proj_nested_global_empty");
+    const nestedTeamProjDirEmptyBaseProj = path.join(testDir, testAppNm + "_proj_nested_global_empty_proj");
     let origDir: string;
 
     const envHost = testEnvPrefix + "_OPT_HOST";
@@ -1347,6 +1349,18 @@ describe("TeamConfig ProfileInfo tests", () => {
                 defaultBaseProfileName: "base_glob",
                 propsToStore: ["areBirdsReal"], sessCfg: { "areBirdsReal": true }, setSecure: undefined,
             });
+        });
+        it("should not add new base to nested project config profile when updating secure creds", async () => {
+            process.env[testEnvPrefix + "_CLI_HOME"] = nestedTeamProjDirEmptyBase;
+            const profInfo = createNewProfInfo(nestedTeamProjDirEmptyBase);
+            await profInfo.readProfilesFromDisk({ projectDir: nestedTeamProjDirEmptyBaseProj});
+            const profiles = profInfo.getAllProfiles();
+            const desiredProfile = "lpar1.zosmf";
+            const upd = { profileName: "lpar1.zosmf", profileType: "zosmf" };
+            await profInfo.updateProperty({ ...upd, property: "user", value: "testxyz", setSecure: true });
+            await profInfo.updateProperty({ ...upd, property: "password", value: "testabc", setSecure: true });
+            const profAttrs = profiles.find(p => p.profName === desiredProfile);
+            expect(!profiles.find(p => p.profName === "lpar1.base"));
         });
     });